beta_version/code/requirements.txt
# Core dependencies
requests==2.31.0
aiohttp==3.9.1
asyncio-mqtt==0.11.1
websockets==12.0
python-dotenv==1.0.0

# Social Media APIs
tweepy==4.14.0
praw==7.7.1
discord.py==2.3.2
telethon==1.30.3
TikTokApi==5.3.5

# Blockchain and Trading
solana==0.30.2
solders==0.18.1
anchorpy==0.19.1
construct==2.10.70

# AI and Data Analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2

# Web Interface
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0

# Database
sqlite3

# Data Processing
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Configuration and Security
cryptography==41.0.8
python-jose==3.3.0
passlib==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
